Move map method to its own controller
